[dingo-server] hotfix bugs about Jackson parser

+++ dingo-server/coordinator/src/main/java/io/dingodb/server/coordinator/config/RaftConfiguration.java
@@ -26,7 +26,7 @@ import lombok.ToString;
 @ToString
 public class RaftConfiguration {
     private String group = "COORDINATOR_RAFT";
-    private int port;
+    private Long port;
     private String initRaftSvrList;
     private NodeOptions node;
 }
